Relax constraints
The caret constraint is a bit too strict.
We can make things a bit easier on consumers if we specify a range.

The PS community is pretty good about specifying breaking changes.
However, most of the version changes between packages
don't actual cause breakage in other people's packages.
Hopefully the community will continue being rigorous about their changes,
and we can rely on that to influence which versions we actually support.