Lock down upper bounds better
We weren't using quite the right syntax for what we wanted.

If we say `x - y`, then we allow all versions in the range including the
bounds. This means that everything broke for the new versions at the
upper bounds.

Instead, we now say `>= x < y`. This specifies that we want all versions
in the range, excluding the upper bound only. This was the original
intent, we just didn't do the right thing.