Bump all npm dependencies to non-breaking versions
We bump everything except `purescript` to the latest version.

There's a huge breaking change in `purescript` as of 0.14.0. Instead of
updating everything in one go, we at least bump up to the latest
non-breaking version.

We'll bump up to a newer version after this. But for now, we want as
many dependencies up to date as possible.