Get dependencies from `spago`
We can't really update any dependencies with `bower` these days.

Given that ecosystem is dealing with a bunch of breaking changes due to
0.14.0, and the response to the breaking changes is to also change how
dependencies are resolved, we jump on the bandwagon and abandon `bower`.

We still don't change the build so we're reliant on `spago` for actually
building, since it's one extra layer we don't really care to deal with.